Moves create_square() into lib.rs
