+++ code/examples/kmeans.rs
@@ -1,3 +1,4 @@
+use code::*;
 // ANCHOR: libraries
 // Import the linfa prelude and KMeans algorithm
 use linfa::prelude::*;
@@ -30,7 +31,7 @@ fn main() {
 
     // ANCHOR: create_model
     let dataset = DatasetBase::from(combined_squares);
-    let rng = rand::thread_rng(); // Random number generator
+    let rng = thread_rng(); // Random number generator
     let n_clusters = 3;
     let model = KMeans::params_with_rng(n_clusters, rng)
         .max_n_iterations(200)
@@ -118,19 +119,4 @@ fn check_array_for_plotting(arr: &Array2<f32>) -> bool {
     } else {
         true
     }
-}
-
-// Creates random points contained in an approximately square area
-pub fn create_square(center_point: [f32; 2], edge_length: f32, num_points: usize) -> Array2<f32> {
-    // We
-    let mut data: Array2<f32> = Array2::zeros((num_points, 2));
-    let mut rng = rand::thread_rng();
-    for i in 0..num_points {
-        let x = rng.gen_range(-edge_length * 0.5, edge_length * 0.5); // generates a float between 0 and 1
-        let y = rng.gen_range(-edge_length * 0.5, edge_length * 0.5);
-        data[[i, 0]] = center_point[0] + x;
-        data[[i, 1]] = center_point[1] + y;
-    }
-
-    data
 }
